requirements.txt
streamlit
transformers
sentence-transformers
faiss-cpu
tqdm
numpy
pandas
python-dotenv
openai
langchain
requests
nltk
pdfplumber
gunicorn
